# BackEnd/tshirt-store-api/week-2/openapi.yaml
openapi: 3.0.3
info:
  title: T-Shirt Store API
  description: |
    REST API for the T-Shirt Store capstone project.
    Supports three roles: Manager, Client, and Delivery Person.
  version: 1.0.0
  contact:
    name: Adrian
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable parameters ────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    AddressIdParam:
      name: addressId
      in: path
      required: true
      schema:
        type: integer
        format: int64

  # ─── Reusable schemas ───────────────────────────────────────────
  schemas:
...
    SignUpRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: Secret123!
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        phone:
          type: string
          example: "+1234567890"
    SignInRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIs...
        user:
          $ref: "#/components/schemas/UserSummary"
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [manager, client, delivery_person]
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          format: password
          minLength: 8

    # ── Category ──
...
    Address:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
          nullable: true
          example: Casa
        recipientName:
          type: string
          example: Demo User
        recipientPhone:
          type: string
          example: "+1234567890"
        line1:
          type: string
          example: 123 Main St
        line2:
          type: string
          nullable: true
        city:
          type: string
          example: San Francisco
        stateRegion:
          type: string
          nullable: true
          example: CA
        postalCode:
          type: string
          nullable: true
          example: "94102"
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
          example: US
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAddressRequest:
      type: object
      required: [recipientName, recipientPhone, line1, city, countryCode]
      properties:
        label:
          type: string
          maxLength: 50
        recipientName:
          type: string
          maxLength: 200
        recipientPhone:
          type: string
          maxLength: 30
        line1:
          type: string
          maxLength: 255
        line2:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 100
        stateRegion:
          type: string
          maxLength: 100
        postalCode:
          type: string
          maxLength: 20
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
        isDefault:
          type: boolean
          default: false
    UpdateAddressRequest:
      type: object
      properties:
        label:
          type: string
          maxLength: 50
        recipientName:
          type: string
          maxLength: 200
        recipientPhone:
          type: string
          maxLength: 30
        line1:
          type: string
          maxLength: 255
        line2:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 100
        stateRegion:
          type: string
          maxLength: 100
        postalCode:
          type: string
          maxLength: 20
        countryCode:
          type: string
          minLength: 2
          maxLength: 2
        isDefault:
          type: boolean

    # ── Product ──
...
    Cart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        totalAmount:
          type: number
          format: float
    CartItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productVariantId:
          type: integer
          format: int64
        productName:
          type: string
        skuCode:
          type: string
        sizeName:
          type: string
        colorName:
          type: string
        imageUrl:
          type: string
          format: uri
          nullable: true
        unitPrice:
          type: number
          format: float
        quantity:
          type: integer
        lineTotal:
          type: number
          format: float
    AddCartItemRequest:
      type: object
      required: [productVariantId, quantity]
      properties:
        productVariantId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1

    # ── Order ──
    OrderSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNumber:
          type: string
        orderStatus:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        subtotal:
          type: number
          format: float
        discountAmount:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        paymentMethod:
          type: string
          enum: [payment_link, payment_intent]
          nullable: true
        createdAt:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/OrderSummary"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            promoCode:
              type: string
              nullable: true
            promoDiscount:
              type: number
              format: float
              nullable: true
            shippingAddress:
              type: object
              properties:
                recipientName:
                  type: string
                recipientPhone:
                  type: string
                line1:
                  type: string
                line2:
                  type: string
                  nullable: true
                city:
                  type: string
                stateRegion:
                  type: string
                  nullable: true
                postalCode:
                  type: string
                countryCode:
                  type: string
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/StatusHistoryEntry"
    OrderItem:
      type: object
      properties:
        productName:
          type: string
        skuCode:
          type: string
        sizeName:
          type: string
        colorName:
          type: string
        imageUrl:
          type: string
          format: uri
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
          format: float
        lineTotal:
          type: number
          format: float
    StatusHistoryEntry:
      type: object
      properties:
        orderStatus:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        changedBy:
          type: string
          nullable: true
          description: User email or "system"
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [addressId]
      properties:
        addressId:
          type: integer
          format: int64
        promoCode:
          type: string
          description: Optional promo code to apply
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            totalItems:
              type: integer
            totalPages:
              type: integer

    # ── Error ──
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
      example:
        statusCode: 400
        message: Validation failed
        error: Bad Request

# ─── Paths ──────────────────────────────────────────────────────────
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new client account
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signin:
    post:
      tags: [Auth]
      summary: Sign in with email and password
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signout:
    post:
      tags: [Auth]
      summary: Sign out (invalidate token)
      operationId: signOut
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Signed out
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: If the email exists, a reset link was sent
        "429":
          description: Too many requests (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successfully (email notification sent)
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Categories ────────────────────────────────────────────────────
...
  /addresses:
    get:
      tags: [Addresses]
      summary: List current user addresses
      operationId: listAddresses
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of current user addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Addresses]
      summary: Create an address for the current user
      operationId: createAddress
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressRequest"
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /addresses/{addressId}:
    patch:
      tags: [Addresses]
      summary: Update one of the current user addresses
      operationId: updateAddress
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AddressIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddressRequest"
      responses:
        "200":
          description: Address updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Address not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Addresses]
      summary: Delete one of the current user addresses
      operationId: deleteAddress
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AddressIdParam"
      responses:
        "204":
          description: Address deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Address not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Products ──────────────────────────────────────────────────────
...
  /cart:
    get:
      tags: [Cart]
      summary: Get current cart (client only)
      operationId: getCart
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current active cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart (client only)
      operationId: addCartItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "201":
          description: Item added to cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cart/items/{itemId}:
    patch:
      tags: [Cart]
      summary: Update cart item quantity (client only)
      operationId: updateCartItem
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Cart item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Cart]
      summary: Remove item from cart (client only)
      operationId: removeCartItem
      security:
        - BearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Item removed
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Orders ────────────────────────────────────────────────────────
  /orders:
    post:
      tags: [Orders]
      summary: Create an order from cart (client only)
      operationId: createOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created (status pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "400":
          description: Cart is empty or stock unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Orders]
      summary: List orders (client sees own, manager sees all)
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, processing, shipped, delivered, cancelled]
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter orders from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter orders up to this date
        - name: minAmount
          in: query
          schema:
            type: number
            format: float
          description: Minimum total amount
        - name: maxAmount
          in: query
          schema:
            type: number
            format: float
          description: Maximum total amount
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrderSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...